+++ data_files/example_dbs/survival_curve.sql
@@ -1110,7 +1110,7 @@ CREATE TABLE time_of_day
     sequence INTEGER UNIQUE,
     tod      TEXT
         PRIMARY KEY,
-    hours    INTEGER NOT NULL DEFAULT 1,
+    hours    REAL NOT NULL DEFAULT 1,
     notes    TEXT
 );
 INSERT INTO "time_of_day" (sequence, tod, hours) VALUES(0,'d',24);
