+++ data_files/temoa_schema_v4.sql
@@ -927,7 +927,7 @@ CREATE TABLE IF NOT EXISTS time_of_day
     sequence INTEGER UNIQUE,
     tod      TEXT
         PRIMARY KEY,
-    hours    INTEGER NOT NULL DEFAULT 1,
+    hours    REAL NOT NULL DEFAULT 1,
     notes    TEXT
 );
 CREATE TABLE IF NOT EXISTS time_period
